% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_CacheEntryList.R
\name{as_CacheEntryList}
\alias{as_CacheEntryList}
\title{Read from list}
\usage{
as_CacheEntryList(l)
}
\arguments{
\item{l}{(\code{list}) The list to convert to a CacheEntryList.}
}
\value{
(\code{CacheEntryList}) The CacheEntryList object created from the list.
}
\description{
Read from list
}
