% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_list.R
\name{as_list}
\alias{as_list}
\title{As list}
\usage{
as_list(x)
}
\arguments{
\item{x}{(\code{CacheMeta | CacheEntry | CacheEntryList}) The object to convert to a list.}
}
\value{
(\code{list}) The list representation of the CacheEntryList.
}
\description{
As list
}
