% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_get}
\alias{cache_get}
\title{Get certain families from CacheEntryList}
\usage{
cache_get(x, families = NULL, quiet = TRUE)
}
\arguments{
\item{x}{(\code{CacheEntryList}) The CacheEntryList object to query.}

\item{families}{(\verb{character vector}) The family names to retrieve.}

\item{quiet}{(\code{logical(1)}) If TRUE, suppress informational messages (default: TRUE).}
}
\value{
(\code{list}) A list of CacheEntry objects matching the specified families. If no families
}
\description{
Get certain families from CacheEntryList
}
\seealso{
Other cache: 
\code{\link{cache_clean}()},
\code{\link{cache_get_weights}()},
\code{\link{cache_read}()},
\code{\link{cache_remove}()},
\code{\link{cache_write}()}
}
\concept{cache}
