% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_get_weights}
\alias{cache_get_weights}
\title{Check which weights are available in a cache entry}
\usage{
cache_get_weights(entry, weights)
}
\arguments{
\item{entry}{(\code{CacheEntry}) The CacheEntry object to check.}

\item{weights}{(\code{numeric}) Vector of weights to check for availability.}
}
\value{
(\code{lgl}) Logical vector indicating which weights are cached).
}
\description{
Check which weights are available in a cache entry
}
\seealso{
Other cache: 
\code{\link{cache_clean}()},
\code{\link{cache_get}()},
\code{\link{cache_read}()},
\code{\link{cache_remove}()},
\code{\link{cache_write}()}
}
\concept{cache}
