% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_remove}
\alias{cache_remove}
\title{Delete entry from cache}
\usage{
cache_remove(x, families = NULL, remove_files = TRUE, cache_dir = NULL)
}
\arguments{
\item{x}{(\code{CacheEntryList}) The CacheEntryList object to modify.}

\item{families}{(\code{character | NULL}) The font families to delete. If NULL, all entries are deleted.}

\item{remove_files}{(\code{logical(1)}) If \code{TRUE} attempt to delete files referenced by removed entries (default: TRUE).}

\item{cache_dir}{(\code{character(1) | NULL}) The cache directory to delete from. If NULL, the default cache directory is used.}
}
\value{
(\code{CacheEntryList}) The modified CacheEntryList with the specified entries removed.
}
\description{
Delete entry from cache
}
\seealso{
Other cache: 
\code{\link{cache_clean}()},
\code{\link{cache_get}()},
\code{\link{cache_get_weights}()},
\code{\link{cache_read}()},
\code{\link{cache_write}()}
}
\concept{cache}
