% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_set}
\alias{cache_set}
\title{Set cache entries}
\usage{
cache_set(x, family, meta)
}
\arguments{
\item{x}{(\code{CacheEntryList}) The CacheEntryList object to modify.}

\item{family}{(\code{character(1)}) The font family to modify.}

\item{meta}{(\code{CacheMeta}) The new metadata to set.}
}
\value{
(\code{CacheEntryList}) The modified CacheEntryList with the updated entry.
}
\description{
Set cache entries
}
