% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_ttf_path.R
\name{cache_ttf_filename}
\alias{cache_ttf_filename}
\title{Compose canonical filename for a cached TTF}
\usage{
cache_ttf_filename(source, font_id, subset, weight, style)
}
\arguments{
\item{source}{(\code{character(1)}) Provider source identifier.}

\item{font_id}{(\code{character(1)}) Font id used to create a filesystem-safe filename.}

\item{subset}{(\code{character(1)}) Glyph subset identifier.}

\item{weight}{(\code{integer(1)}) Font weight.}

\item{style}{(\code{character(1)}) Style string (e.g. "normal", "italic").}
}
\value{
(\code{character(1)}) Filename (not including the cache directory) for the cached TTF.
}
\description{
Compose canonical filename for a cached TTF
}
