% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_ttf_path.R
\name{cache_ttf_path}
\alias{cache_ttf_path}
\title{Compute canonical cache path for a TTF file}
\usage{
cache_ttf_path(source, font_id, subset, weight, style, cache_dir = NULL)
}
\arguments{
\item{source}{(\code{character(1)}) Provider source identifier.}

\item{font_id}{(\code{character(1)}) Font id used for filenames.}

\item{subset}{(\code{character(1)}) Glyph subset identifier.}

\item{weight}{(\code{integer(1)}) Font weight.}

\item{style}{(\code{character(1)}) Style string (e.g. "normal", "italic").}

\item{cache_dir}{(\code{character | NULL}) Cache directory to use (default: NULL)}
}
\value{
(\code{character(1)}) Path to the cached \code{.ttf} file.
}
\description{
Compute canonical cache path for a TTF file
}
