% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_write}
\alias{cache_write}
\title{Write CacheEntryList to disk as JSON}
\usage{
cache_write(x, cache_dir = NULL, quiet = TRUE)
}
\arguments{
\item{x}{(\code{CacheEntryList}) The CacheEntryList object to write to disk.}

\item{cache_dir}{(\code{character(1)}) The cache directory to write to (default: NULL).}

\item{quiet}{(\code{logical(1)}) Whether to suppress output messages (default: TRUE).}
}
\value{
(\code{NULL}) Invisibly returns NULL.
}
\description{
Write CacheEntryList to disk as JSON
}
\seealso{
Other cache: 
\code{\link{cache_clean}()},
\code{\link{cache_get}()},
\code{\link{cache_get_weights}()},
\code{\link{cache_read}()},
\code{\link{cache_remove}()}
}
\concept{cache}
