% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheEntryList.R
\name{CacheEntryList}
\alias{CacheEntryList}
\title{S7 list of cache entries (CacheEntryList)}
\usage{
CacheEntryList(entries = list())
}
\arguments{
\item{entries}{(\verb{list(1+)}) A non-empty list of \code{CacheEntry} objects.}
}
\value{
(\code{S7_object}) A validated S7 \code{CacheEntryList} object.
}
\description{
S7 list of cache entries (CacheEntryList)
}
