% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_variant_generic.R
\name{download_variant_generic}
\alias{download_variant_generic}
\title{Download and (if needed) convert a provider artifact to a local TTF  file for a given family/weight/style and return the local path.}
\usage{
download_variant_generic(
  provider,
  family,
  weight,
  style,
  subset = "latin",
  cache_dir = NULL,
  quiet = FALSE
)
}
\arguments{
\item{provider}{(\code{FontProvider}) Provider object with url_template and source.}

\item{family}{(\code{character(1)}) Family identifier.}

\item{weight}{(\code{integer(1)}) Font weight to fetch (100-900).}

\item{style}{(\code{character(1)}) Style (e.g. "normal", "italic").}

\item{subset}{(\code{character(1)}) Glyph subset to request (default: "latin")}

\item{cache_dir}{(\code{character | NULL}) Cache directory to use (default: NULL)}

\item{quiet}{(\code{logical(1)}) Suppress warnings/messages (default: FALSE)}
}
\value{
(\code{character | NULL}) Path to the local \code{.ttf} file on success, or \code{NULL} on failure.
}
\description{
Download and (if needed) convert a provider artifact to a local TTF  file for a given family/weight/style and return the local path.
}
