% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cache_dir}
\alias{get_cache_dir}
\title{Get the package cache directory for fonts}
\usage{
get_cache_dir()
}
\value{
(\code{character(1)}) Absolute path to the cache directory.
}
\description{
Determine a platform-appropriate cache directory for AddFonts and
ensure it exists, creating it when necessary.
}
