% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_font_cache_dir.R
\name{get_font_cache_dir}
\alias{get_font_cache_dir}
\title{Get Font Cache Directory}
\usage{
get_font_cache_dir()
}
\value{
Character string with absolute path to the fonts cache directory.
}
\description{
Returns the directory where fonts are/should be cached.
Uses platform-appropriate cache locations via the rappdirs package.
}
\examples{
\dontrun{
# Get font cache directory
cache_dir <- get_font_cache_dir()
print(cache_dir)
}

}
