% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_provider_details}
\alias{get_provider_details}
\title{Get provider details from internal data}
\usage{
get_provider_details(provider)
}
\arguments{
\item{provider}{(\code{character(1)}) Provider id/name (e.g. "bunny").}
}
\value{
(\code{FontProvider}) A validated FontProvider object.
}
\description{
Load and return a FontProvider object for the specified provider.
The providers data is stored in the package's internal sysdata.rda.
}
