% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_id}
\alias{safe_id}
\title{Create a filesystem-safe id from a name}
\usage{
safe_id(name)
}
\arguments{
\item{name}{(\code{character(1)}) Input name to sanitise.}
}
\value{
(\code{character(1)}) Sanitised identifier.
}
\description{
Replace disallowed characters with \code{-} and convert to lower-case so
the resulting id is safe to use in filenames.
}
