% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_FontProvider.R
\name{as_FontProvider}
\alias{as_FontProvider}
\title{Construct a \code{FontProvider} from a named list}
\usage{
as_FontProvider(x)
}
\arguments{
\item{x}{(\code{list}) Named list (e.g. from JSON) with provider details.}
}
\value{
(\code{FontProvider}) The corresponding \code{FontProvider} object.
}
\description{
Construct a \code{FontProvider} from a named list
}
