% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_read}
\alias{cache_read}
\title{Read cache entry from disk}
\usage{
cache_read(cache_dir)
}
\arguments{
\item{cache_dir}{(\code{character | NULL}) Cache directory to use. Use \code{\link[=get_cache_dir]{get_cache_dir()}} to get the default cache directory.}
}
\value{
(\code{CacheEntryList}) The cache index as a \if{html}{\out{<CacheEntryList>}} if found and valid.
}
\description{
Read cache entry from disk
}
\seealso{
Other cache: 
\code{\link{cache_clean}()},
\code{\link{cache_get}()},
\code{\link{cache_get_weights}()},
\code{\link{cache_remove}()},
\code{\link{cache_write}()}
}
\concept{cache}
