% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FontProvider.R
\name{FontProvider}
\alias{FontProvider}
\title{Font provider specification (FontProvider)}
\usage{
FontProvider(
  source = character(0),
  url_template = character(0),
  conversion = NULL,
  conversion_ext = NULL,
  aliases = list()
)
}
\arguments{
\item{source}{(\code{character(1)}) Provider id/name (e.g. "bunny").}

\item{url_template}{(\code{character(1)}) URL template used to construct download URLs.}

\item{conversion}{(\code{character(1) | NULL}) Optional conversion function name (as string) or \code{NULL}.}

\item{conversion_ext}{(\code{character(1) | NULL}) Original extension handled by the provider (e.g. "woff2").}

\item{aliases}{(\code{list | NULL}) Optional list of alias names to match (e.g. "fonts.bunny.net").}
}
\value{
(\code{FontProviders}) S7 class representing a font provider specification.
}
\description{
Font provider specification (FontProvider)
}
